Tried to do preprocessing standard scaler after splitting into test and train (more in description)
This preprocessing article https://towardsdatascience.com/preprocessing-with-sklearn-a-complete-and-comprehensive-guide-670cb98fcfb9, mentioned that it was important to do any feature scaling (like standard scaler) AFTER splitting into test and train. The problem is that it turns my discrete data into continuous when I do it after splitting into test and train. Either I need to figure out how to turn it to discrete again, or find a different feature scaling algorithm.