More info in description
Ran a RFC using the standard scaled X_train. Just realized that X_train is the actual spectra and y_train is the 1's and 0's. I knew that before but forgot, so that's why my last commit (tried to std scale the y's) did not work.

Anyway, it got me an error and really bad scores. Wanted to commit it just to save a note about this.