Plotted my false negatives using ipywidget
Plotted highz/spec_err and highz and spec_err in the same subplot (array?). I'm using ipywidget so I can just press the arrow key and look through the false_negatives/positives etc etc quickly instead of having to retype and index and re-running. Doing this to see if the ML algorithm actually achieved the 98.55% accuracy we think it got. Visual inspection!