Cleaning up GitHub, moving files.
I moved Valentina's code that I modified to secondary research. I did not want to delete because maybe one day I would need it or it could be helpful to others, but I'm not using it cause we realized early in the Spring 2023 semester that for my high-z classification we did not need an autoencoder and we did not need t-SNE to get high-z sources from me. The idea with the high-z sources from t-SNE was to have a sample of random high-z sources, but we had to encode and decode and we felt it best to just use the entire noise/highz spectrums without cutting and without encoding/decoding.